добавляет FilterView (фильтр точек маршрута)
