+++ public/index.html
@@ -22,6 +22,8 @@
             <div class="trip-controls__filters">
               <h2 class="visually-hidden">Filter events</h2>
               <!-- Фильтры -->
+                <filter-view ></filter-view>
+
             </div>
           </div>
