+++ src/main.js
@@ -1,4 +1,5 @@
 import './views/brief-view.js';
+import './views/filter-view.js';
 
 /**
  * @type {import('./views/brief-view').default}
@@ -6,4 +7,11 @@ import './views/brief-view.js';
 
 const BriefView = document.querySelector('brief-view');
 
+/**
+ * @type {import('./views/filter-view').default}
+ */
+
+const FilterView = document.querySelector('filter-view');
+
 BriefView.render();
+FilterView.render();
