+++ src/views/point-edit-view-old.js
