+++ src/views/trip-info-view-old.js
